reddit/requirements.txt
boto3==1.26.102
botocore==1.29.102
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
jmespath==1.0.1
praw==7.7.0
prawcore==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
wordfilter==0.2.7
